Cache more file patterns
The caching file pattern factory was not passed around everywhere
it should be, resulting in many patterns being created from scratch.
Also the normal factory will now always cache the default patterns,
since they are always needed and very expensive.

This fix is more of a band-aid. We should re-evaluate if this cache
is necessary or if we can make our String->Spec conversion cheaper.
Also if we can't make it cheaper, we should try to remove the spec
factory from the PatternSet API and instead pass it in as a parameter
when we actually want to walk the tree.